Be able to set haproxy service to disabled (enable=False) but prevent salt run from stopping haproxy (with service.dead)
This is required to use this formula when haproxy is managed via cluster management (like Pacemaker)

Service isn't allowed to be enabled, because clustermanager manages stopping and starting of haproxy.

On the other hand service.dead would stop haproxy on Salt run, even if it is already started by clustermanager, which could cause harm.